alert delete

+++ src/app/components/cus-profile/cus-profile.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ApiService } from '../../services/api.service';
 import { Customer } from '../../models/customer';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cus-profile',
@@ -14,7 +15,11 @@ export class CusProfileComponent implements OnInit {
   cusEdit: Customer;
   cusData: any;
   cus: any;
-  edit: any;
+
+  moo: any;
+  soi: any;
+  road: any;
+  address: any;
 
   constructor(
     public activatedRoute: ActivatedRoute,
@@ -27,10 +32,31 @@ export class CusProfileComponent implements OnInit {
   ngOnInit(): void {
     this.id = this.activatedRoute.snapshot.params["id"];
     this.apiService.getCustomer(this.id).then((res: any) => {
-      this.edit = res;
-      this.cusData = res;
+      this.getData(res);
+    
+      if (res[0].moo == null) {
+        this.moo=""
+      } else {
+        this.moo = " ม."+res[0].moo;
+      }
+      if (res[0].soi == null || res[0].soi == " ") {
+        this.soi=""
+      }else {
+        this.soi = "ซอย"+res[0].soi;
+      }
+      if (res[0].road == null || res[0].road == " ") {
+        this.road=""
+      }else {
+        this.road = "ถนน"+res[0].road;
+      }
+      this.address = res[0].number + this.moo + " " +this.soi + " " +this.road + " " + res[0].sub_district + " " + res[0].district + " " + res[0].province + " " + res[0].postal_code;
       this.cusEdit = res[0]
+
     });
+  }
+
+  getData(data: any) {
+    this.cusData=data;
   }
 
   update() {
@@ -51,12 +77,23 @@ export class CusProfileComponent implements OnInit {
   }
 
   delete() {
-    //Delete item in Student data
-    this.apiService.deleteCustomer(this.id).then((res: any) => {
-      // console.log('deleted '+ this.id);
-      this.router.navigate(['customer']);
-      this.refreshCustomers();
-    });
+    Swal.fire({
+      title: 'ยืนยันการลบข้อมูลลูกค้า',
+      text: this.cusData[0].cus_fullname,
+      showDenyButton: true,
+      confirmButtonText: `ใช่`,
+      denyButtonText: `ไม่ใช่`,
+      icon: "warning",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.apiService.deleteCustomer(this.id).then((res: any) => {
+          // console.log('deleted '+ this.id);
+          this.router.navigate(['customer']);
+          this.refreshCustomers();
+        });
+        Swal.fire('ลบสำเร็จ!', '', 'success')
+      }
+    })
   }
 
   refreshCustomers() {
